myimages/requirements.txt
asgiref==3.5.0
Django==3.2.16
django-imagekit==4.1.0
django-session-timeout==0.1.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
